# Additional numba optimizations