--- a/galois/conway.py
+++ b/galois/conway.py
@@ -34,7 +34,7 @@ class ConwayDatabase:
         result = self.cursor.fetchone()
 
         if result is None:
-            raise LookupError(f"Frank Luebeck's database of Conway polynomials doesn't contain an entry for GF({characteristic}^{degree})")
+            raise LookupError(f"Frank Luebeck's database of Conway polynomials doesn't contain an entry for GF({characteristic}^{degree}). See here http://www.math.rwth-aachen.de/~Frank.Luebeck/data/ConwayPol/index.html for his complete list of polynomials.")
 
         coeffs = result[0]
         coeffs = list(map(int, coeffs[1:-1].split(",")))  # List of degree-ascending coefficients