… .github/workflows/release.yml
… README.md
… docs/_templates/module.rst
… galois/meta_gfp.py
… galois/meta_gfpm.py
… galois/meta_mixin_ufunc.py
… galois/poly.py
… galois/prime.py
… setup.cfg
… tests/fields/test_arithmetic_exceptions.py
… tests/fields/test_numpy_ufuncs.py
… tests/polys/test_function.py
… tests/test_algorithms.py
… tests/test_factor.py
… tests/test_math.py
… tests/test_primes.py