This pull request includes many performance speed-ups, mostly optimizing how numba is used.