@mhostetter 2021-05-05
Use native numpy bitwise ufuncs to achieve GF(2) arithmetic instead of custom ufuncs.

Before
```python
In [1]: import galois                                                                 

In [2]: import numpy as np                                                            

In [3]: a = galois.GF2.Random(100_000)                                                

In [4]: b = galois.GF2.Random(100_000, low=1)                                         

In [5]: %timeit a + b                                                                 
166 µs ± 3.68 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [6]: %timeit a * b                                                                 
173 µs ± 9.36 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [7]: %timeit a / b                                                                 
361 µs ± 1.06 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [8]: %timeit np.negative(a)                                                        
134 µs ± 7.41 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [9]: %timeit np.reciprocal(b)                                                      
141 µs ± 10.5 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [10]: %timeit a ** 100                                                             
299 µs ± 9.12 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [11]: %timeit a ** 2                                                               
237 µs ± 50 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [13]: %timeit np.log(b)                                                            
232 µs ± 1.52 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

After
```python
In [1]: import galois                                                                 

In [2]: import numpy as np                                                            

In [3]: a = galois.GF2.Random(100_000)                                                

In [4]: b = galois.GF2.Random(100_000, low=1)                                         

In [5]: %timeit a + b                                                                 
55.4 µs ± 1.53 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [6]: %timeit a * b                                                                 
57.8 µs ± 1.76 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [7]: %timeit a / b                                                                 
246 µs ± 2.34 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [8]: %timeit np.negative(a)                                                        
2.07 µs ± 80.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [10]: %timeit np.reciprocal(b)                                                     
193 µs ± 5.18 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [11]: %timeit a ** 100                                                             
291 µs ± 3.15 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [12]: %timeit a ** 2                                                               
2.25 µs ± 124 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [13]: %timeit np.log(a)                                                            
236 µs ± 1.91 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```