@mhostetter 2021-05-05
Use BLAS/LAPACK for prime fields to make matrix multiplication much faster.

Before
```python
In [1]: import galois                                                                               

In [2]: GF = galois.GF2                                                                             

In [3]: A = GF.Random((100,100))                                                                    

In [4]: B = GF.Random((100,100))                                                                    

In [5]: %timeit A @ B                                                                               
1.94 s ± 42.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [2]: GF = galois.GF(galois.next_prime(1_000_000))                                                

In [3]: A = GF.Random((100,100))                                                                    

In [4]: B = GF.Random((100,100))                                                                    

In [5]: %timeit A @ B                                                                               
2.27 s ± 180 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After
```python
In [1]: import galois                                                                               

In [3]: GF = galois.GF2                                                                             

In [4]: A = GF.Random((100,100))                                                                    

In [5]: B = GF.Random((100,100))                                                                    

In [6]: %timeit A @ B                                                                               
727 µs ± 21.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [2]: GF = galois.GF(galois.next_prime(1_000_000))                                                

In [3]: A = GF.Random((100,100))                                                                    

In [4]: B = GF.Random((100,100))                                                                    

In [5]: %timeit A @ B                                                                               
806 µs ± 18.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```