@mhostetter 2021-05-05
Greatly speed up `is_primitive()` by only checking the prime factors of `p^n - 1`.

Before
```python
In [1]: import galois                                                                        

In [2]: p = galois.conway_poly(2, 8)                                                         

In [3]: %timeit galois.is_primitive(p)                                                       
15 s ± 285 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After
```python
In [1]: import galois                                                                        

In [2]: p = galois.conway_poly(2, 8)                                                         

In [3]: %timeit galois.is_primitive(p)                                                       
41.6 ms ± 135 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```