@mhostetter 2021-05-06
Speed up multiplicative inverse of prime fields.

Before
```python
In [1]: import numpy as np                                                                                  

In [2]: import galois                                                                                       

In [3]: GF = galois.GF(galois.next_prime(2**21))                                                            

In [4]: b = GF.Random(100_000, low=1)                                                                       

In [5]: %timeit np.reciprocal(b)                                                                            
74.1 ms ± 4.87 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

After
```python
In [1]: import numpy as np                                                                                  

In [2]: import galois                                                                                       

In [3]: GF = galois.GF(galois.next_prime(2**21))                                                            

In [4]: b = GF.Random(100_000, low=1)                                                                       

In [5]: %timeit np.reciprocal(b)                                                                            
14.1 ms ± 144 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```