@mhostetter 2021-05-06
Move raising arithmetic exception inside the numba ufuncs. This eliminates the need to do costly `np.any()` checks on the inputs before calling the ufunc. Much faster.

Before
```python
In [1]: import galois

In [2]: GF = galois.GF(31)

In [3]: a = GF.Random(1_000_000)

In [4]: b = GF.Random(1_000_000, low=1)

In [5]: %timeit a / b
7.06 ms ± 974 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

After
```python
In [1]: import galois

In [2]: GF = galois.GF(31)

In [6]: a = GF.Random(1_000_000)

In [7]: b = GF.Random(1_000_000, low=1)

In [8]: %timeit a / b
3.04 ms ± 29.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```